requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
requests==2.28.1
python-dotenv==0.19.2
